+++ test/devrev_test.dart
@@ -1,29 +1,38 @@
-import 'package:flutter_test/flutter_test.dart';
 import 'package:devrev/devrev.dart';
-import 'package:devrev/devrev_platform_interface.dart';
-import 'package:devrev/devrev_method_channel.dart';
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-
-class MockDevrevPlatform
-    with MockPlatformInterfaceMixin
-    implements DevrevPlatform {
-
-  @override
-  Future<String?> getPlatformVersion() => Future.value('42');
-}
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  final DevrevPlatform initialPlatform = DevrevPlatform.instance;
+  TestWidgetsFlutterBinding.ensureInitialized();
+  group("$DevRev", () {
+    const MethodChannel channel = MethodChannel("devrev");
 
-  test('$MethodChannelDevrev is the default instance', () {
-    expect(initialPlatform, isInstanceOf<MethodChannelDevrev>());
-  });
+    final List<MethodCall> log = <MethodCall>[];
 
-  test('getPlatformVersion', () async {
-    DevRev devrevPlugin = DevRev();
-    MockDevrevPlatform fakePlatform = MockDevrevPlatform();
-    DevrevPlatform.instance = fakePlatform;
+    late DevRev devrev;
 
-    expect(await devrevPlugin.getPlatformVersion(), '42');
+    setUp(() {
+      channel.setMockMethodCallHandler((MethodCall call) async {
+        log.add(call);
+        if (call.method == "showSupport") {
+          print("PLuG invoked");
+          return null;
+        }
+        return {};
+      });
+
+      devrev = DevRev();
+      log.clear();
+    });
+
+    group("#showSupport", () {
+
+      test('PLuG Test', () async {
+
+        await devrev.showSupport();
+
+        expect(log, <Matcher>[isMethodCall('showSupport', arguments: {})]);
+      });
+    });
   });
 }